+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.59](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.58...@synapsecns/widget@0.0.59) (2024-02-22)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.0.58](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.57...@synapsecns/widget@0.0.58) (2024-02-15)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.0.57](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.56...@synapsecns/widget@0.0.57) (2024-02-15)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.0.56](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.55...@synapsecns/widget@0.0.56) (2024-02-15)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.0.55](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.54...@synapsecns/widget@0.0.55) (2024-02-13)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.0.54](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.53...@synapsecns/widget@0.0.54) (2024-02-13)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
 ## [0.0.53](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.0.52...@synapsecns/widget@0.0.53) (2024-02-12)
 
 **Note:** Version bump only for package @synapsecns/widget